app/requirements.txt
asgiref==3.5.2
Django==4.1.0
sqlparse==0.4.3
psycopg2-binary
tzdata
django-browser-reload
Pillow
djangorestframework
djangorestframework-gis